Improve `docker-compose.yml` commands with guards
We decided to add some sort of guards to `command`-s we run inside some
containers: rails, sidekiq, and webpacker. To be sure that these
commands are available. This reduces number of errors users see in the
output of `docker-compose up`.

Also, we renamed `runner` service to `shell` (just matter of taste).